+++ controllers/index.js
@@ -1,0 +1,5 @@
+const ctrlContacts = require("./contacts");
+
+module.exports = {
+    ...ctrlContacts,
+}
