+++ routes/api/contacts.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const ctrl = require('../../controllers/contacts');
+const ctrl = require('../../controllers');
 
 const router = express.Router();
 
@@ -18,6 +18,6 @@ router.delete('/:contactId', isValidId, ctrl.deleteById);
 
 router.put('/:contactId', isValidId, validates.validateBody(schemas.addSchema), ctrl.updateById);
 
-router.patch('/:contactId/favorite', isValidId, validates.validateFavorite(schemas.updateFavoriteSchema), ctrl.updateStatusContact)
+router.patch('/:contactId/favorite', isValidId, validates.validateFavorite(schemas.updateFavoriteSchema), ctrl.updateStatusContact);
 
 module.exports = router;
